// src/payload-types.ts
export interface Page {
  id: number;
  title: string;
  slug: string;
  layout?:
    | {
        language?: ('html' | 'css' | 'javascript') | null;
        code?: string | null;
        id?: string | null;
        blockName?: string | null;
        blockType: 'codeInjection';
      }[]
    | null;
  meta?: {
    title?: string | null;
    description?: string | null;
  };
  updatedAt: string;
  createdAt: string;
}